Use save_and_reraise_exception() from common.
This patch common.excutils from openstack-common, which includes
save_and_reraise_exception().  The patch also converts the code
base to use it from there instead of cinder.utils and then removes
it from cinder.utils.

The initial motivation for this was removing another cinder dependency
from cinder.rpc so that it can eventually be moved to openstack-common.

Change-Id: I7354ca51a02aec9c709cf33f77d4abc46acc2742