Send 'create volume from snapshot' to the proper host
A simple solution for bug 1008866. When creating volume from snapshot on
multicluster, in volume it will check if snapshot_id is set. If snapshot_id
is set, make the call create volume directly to the volume host where the
snapshot resides instead of passing it through the scheduler. So snapshot can
be copy to new volume.

Change-Id: Ie9c1a77f62abc40e294b1d0c604cf885652728da