storwize-svc: improved test coverage and fixes.
Added test cases to cover more error paths, now providing 97% test
coverage of the storwize-svc driver. Some tests uncovered bugs in the
driver, which are fixed by this patch. In addition, there are some
fixes to the code included which address comments from the submission
of this driver to nova-volume and sync the two versions.

Change-Id: I7c646885a7f2b1e761725c14940f1d741996beb4