add get_location method for images
This is useful for determining whether the backend storage for the
image in Glance is the same as the storage used in a volume driver.

The direct_url is only available in the v2 images API, so add a
version parameter to each request. As more parts of the v2 API are
used, this parameter will become more useful.

Change-Id: I5ca06b6d80ffe4118f8dd381ed11283573bce71f
Signed-off-by: Josh Durgin <josh.durgin@inktank.com>