Handle missing 'provider_location' in rm_export.
Updates the remove_export function in the ISCSIDriver so that
it handles the case where provider_location isn't set on
a volume. This can occur if a volume is created but not exported.

This was a regression from a previous commit
(9785963c841aff5ea1c159da81f93c2bf62c70c6) which changed the way iscsi
targets are created/deleted.

Fixes LP Bug #1046484.

Change-Id: Ie762b092dd2a84fe4f3b72c46c03a97ae67c53b7