Add lun number (0) to model_update in HpSanDriver
The HpSanDriver was not setting the lun number in model_update
and as a result default value of 1 was being used.  Trouble is the
Lun number used by LeftHand is 0, so the connect info would be
wrong and fail when trying to attach the volume.

Fixes bug #1053717

Change-Id: If82a66c6a78ab03da09223a7b79abe559a6f702f
(cherry picked from commit f55198570fd59b7a2a4ddd2f5e77b3d03d0a3d64)