Remove unused code: check_for_export.
This code has been unused since cinder was split from nova.  It used to
be used as a safety check during live migration.  The check was removed
recently in nova, since it wasn't being used with Cinder anyway (just
nova-volumes).

An alternative approach would be to expose this from Cinder somehow if
we feel it's important, but until then, there's no reason to keep dead
code around.

Change-Id: I5adfb405c1bd04bcc1ad61d511afb227a198f4ed