Return volume type name on volume create
Update openstack api controller for volumes to fully inflate the
returned volume model on create to populate the volume_type fields so
the response serializer will return the name of the volume_type instead
of the id on volume create requests.  Returning the name on create is
consistent with the show and index commands.

fixes bug #1071436

Change-Id: Ie091cf824a2df40211d8bcfa7f177fc851d737f5