Remove dm_setup(remove) call in volume_delete.
While it makes sense to perform the dm_setup(remove) before
performing the lvm remove, it seems to cause some buffer IO
errors as evident by kern.log.

I'm also a bit suspicous in that it seems the delete timeouts
we see seems to have increased in frequency with this addition.
I propose this step is removed until we completely understand the
impact of the kernel error on Precise kernels.

Fixes bug: #1075675

Change-Id: I7ae4decea42d8f070d774d3e6376686b3e5b4f83