Add hosts extension to Cinder.
This adds the hosts extension to Cinder which is similar to
the extension by the same name in Nova.

This extension allows an admin user to check the status of
Cinder services on different hosts.  In addition it also
provides the capability to get host and project specific information
regarding volume count and space allocated for volumes on a specified
host.

Implements blueprint cinder-hosts-extension

Change-Id: Ic3679cfcd900baecdc2da17e42194cac99c9d422