Add CONTRIBUTING file.
If the CONTRIBUTING[.md] file exists, Github will show a link to it to
anyone who files an issue on Github or opens a pull request on
Github. This files points folks to the OpenStack wiki page which
provides the proper instructions on how to contribute to OpenStack.

See also <https://github.com/blog/1184-contributing-guidelines>.

Change-Id: I8fef06b71aa993c4c68a1ec92c39a540318aaf31