Add get_cluster_stats to SolidFire driver
With the filter scheduler addition we added get_stats to the
drivers.  This provides the required data for the filtering
mechanisms and is called by the scheduler.

The change adds the required call in the driver and also adds
an update routine to it's init to populate this info on startup.
This also adds the fake api response so that the tests can perform
the init routine and get the data, while giving the added bonus of
testing the new functions at the same time.

Change-Id: Ia75e078f8e73883131cb1e68f6e6713fce1ab518