Fix create volume from image.
Fix for bug 1129199

During creating a volume from an image, the volume status changed
from creating to active, skipping the download phase. This bug was most
likely introduced during:
https://review.openstack.org/20514
This patch fixes the issue by change image_location back to image_id at
the correct place.
This patch also fixes an error in the exception handling, and the update
of the volume status after creation.

Change-Id: Ib9b647031c029387b776c27f9233388bd2c67e9c