NetApp: Fix race condition in 7-mode iSCSI driver with DFM.
A race condition was discovered in the driver while creating
or deleting multiple volumes in a very short interval of time
(can be reproduced by making successive API calls to Cinder).
This fix decorates the methods responsible for race condition
with a synchronized() decorator. The problem was with the
management software: DFM, and in principle, there should not
be two processes trying to modify the DFM dataset object at
the same time. This patch prevents from such a possibility.

Fixes bug 1091480

Change-Id: I11cfc8868171acec356f4140aba52b8078d109df