Improve logging for volume operations via manager
A large percentage of the logging information was
being done using debug.  This change moves some
of the logging information to be info level and
also tries to add some consistency in terms of
the general create/delete operations.

This includes moving logging for the basic operations
into the manager and omitting the repetive messages from
the drivers.

Fixes bug: 1133131

Change-Id: I792efde67f30d637431ad2fa45714e50bf9c705d