Add some useful log to filter scheduler.
When the filter scheduler is unable to find a suitable host to
deploy a volume on it simply logged a warning "not hosts found".
This isn't very helpful, so this patch adds some logging info as
far as why no valid host was available.  This just focuses on the
capacity filtering as that's where I've commonly seen this issue
particularly in tempest runs.

Also noticed that the get status was checking for "if Not capacity",
but this is wrong because free capacity==0 would be interpretted as
failing to get capacity.  Change this to "if is None".

Fixes bug: 1136147

Change-Id: Ia1ce508826ee69ae68a250ea68616909c59974e5