swift backup service checks version during restore
Modified swift backup service to check metadata version during restore
and raise an error if the backup version isn't a version that the
service knows how to handle. The versions which can be handled are
described in a dictionary mapping versions to methods which can handle
them. This will facilitate graceful handling of newer backup formats by
the swift backup service when we introduce changes.

Fixes bug: 1136174

Change-Id: Id7d05848fd448ce21f641e5cd6945477702cbe38