Fix 3PAR driver hiding existing host error
This fixes a problem with the 3par drivers where creating
a 3par host was failing but there was no message.  The driver
would fail to find the host after the creation failed and the
log entry would show that the host didn't exist.   The real
error was failing to create the host.

bug #1154700

Change-Id: I2267d958f10cc73c6c5f34d224ea3452a4402f18