Speedup solidfire unit tests
The SolidFire driver has a sleep that's used in a retry loop.  This
sleep was being called every iteration of the loop though
(instead of just when the call failed).

This made SolidFire the slowest test in the unit tests, and obviously
it's just stupid anyway.  So this change fixes that and only performs
the sleep if the call actually failed.

Change-Id: I560f2fdb6bf16a6a9b26184234225349ab62e816