Remove unused tools/rfc.sh
This has been deprecated in favour of git-review for quite a while now
and was removed from Nova 10 months ago.

Backport of git commit ae4aecdc7e554f78156c91c7c0b2436638bec5eb in
master.

This is a zero risk commit, as it removes something from tools that
isn't used. It also means removing the only GPL licensed file in the
whole of the cinder repository for 2013.1.1 release.

Change-Id: Ic4bfa1aee3a4b6a7588be3678ac033a1b13e2b18