Use a SSH pool to manage SSH connection
Use a SSH pool to hold all SSH clients.It allows 4 SSH clients
at most to connect to the SSH server at the same time.
This patchset also enables every SSH client connect to the other
controller when they failed to connect to the current controller.
For example,failed to A,then to B,or,failed to B,then to A.

Fixes bug: 1162251
Change-Id: I86f7f684639034be97ddf2031e61ac6bf3a196ad