Fix backup manager formatting error.
ID was being printed as %d instead of %s, causing tests to fail. This
error was previously undetected.
In addition, NetApp tests are failing due to a bug in the suds
package, which is no longer actively maintained. This patch adds skips
to the problematic tests, which should be removed once the issue is
resolved.

Change-Id: I5645909f59aa2ea3dc536ed4707ea75754307050
Fixes: bug 1166600