NetApp: Fix failing NetApp tests
The tests were failing due to an upgrade in the testing framework
(possibly due to upgrade of nose package from 1.2.1 to 1.3.0).
The actual reason was a bug in the suds client. A simple
workaround is utilised here, which sets the logging level of a
suds file to INFO instead of DEBUG.

Some 'FutureWarning' warnings from the test output were also
fixed.

Change-Id: Ia67476c23bffa91264cf79f6a26944b450c4f4fc