Deprecating old dot path locations for Folsom configs
In grizzly, we made a lot of structure changes. This deprecates the old
locations so the dot paths in old configs still work but give warnings
in the logs of what needs to be updated.

Fixes: bug #1168573
Change-Id: I7b1754b0e59a923afa623871fc3be07362bf1bdc
(cherry picked from commit c355d68a868e9232a80c2f22557ffe228fcb4f8f)