Clear volumes stuck in 'downloading'
If the CinderVolume service is restarted while an image-copy is
happening, the volume gets left in a 'downloading' state and
cannot be used or deleted. This fix adds code to init_host to
look for volumes (on this host) in such a state, and move them
to an error state. It also calls clear_download in the driver
in case the driver needs to perform an action, such as
detaching the volume.

Fixes Bug #1172645

Change-Id: Ia5dab9ebf997f573c0b4e57d983f5996af81ede5