Fix LHN driver to allow backend name configuration
The LHN driver wasn't allowing custom volume_backend_name to be set
via multi-backend configuration input.  In addition there were some
issues with the updating that are also addressed in this patch.

There are other drivers that are going to need updated/fixed for
this same problem, but those will be addressed in a separate patch/bug.

Fixes bug: 1173037

Change-Id: Iae247a500739d02e145511ebe96dddaff8966419